refactored out controller features into their own file
this included making many of the function asynchronous
and also adding some code to the energy policy file
