+++ apps/libsignpost/signpost_controller.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+int signpost_controller_init (void);
+
+#ifdef __cplusplus
+}
+#endif
